Make campaign calendars configurable
You can now choose from a list of predefined calendars when creating a
campaign. For now this includes Harptos, Odreian, and Tide.

Odreian is the default at the moment as I don't want this PR to affect
existing campaigns. Because there aren't very many at the moment we
could just manually add the calendar property to each of them and then
set the default to Harptos.

I've added a version number to each of the calendar IDs so that we can
iterate on them later if we want to without old campaign dates being
knocked out.

Matt: I expect that there's loads wrong with this because I'm not at all
familiar with Meteor or React.